backend/requirements.txt
Django~=5.0
psycopg2-binary~=2.9.9
djangorestframework~=3.14
python-dotenv~=1.0.0
dj_database_url~=2.1.0
djangorestframework-simplejwt~=5.3.1
pyjwt~=2.8.0
certifi~=2023.11.17
charset-normalizer~=3.3.2
idna~=3.6
requests~=2.31.0
urllib3~=2.1.0
Pillow~=10.2.0
django-cors-headers~=4.3.1